+++ src/model/endpoint.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct Endpoint {
+    pub(crate) path: String,
+    pub(crate) method: String,
+    pub(crate) groups: Vec<String>,
+}
