+++ src/model/mod.rs
@@ -1,0 +1,14 @@
+mod auth_obj;
+mod config;
+mod domain;
+mod endpoint;
+mod jwt_payload;
+
+pub type AuthObj = auth_obj::AuthObj;
+pub type Config = config::Config;
+pub type ConfigError = config::ConfigError;
+pub type Domain = domain::Domain;
+pub type Endpoint = endpoint::Endpoint;
+pub type JwtPayload = jwt_payload::JwtPayload;
+
+use super::util;
