+++ src/util/mod.rs
@@ -1,0 +1,2 @@
+#[macro_use]
+pub mod load_json;
